story_downloader/requirements.txt
PyQt6>=6.5
requests
beautifulsoup4
python-docx
